src/speed.worker/requirements.txt
nameko==2.11.0
python-dotenv==0.9.1
influxdb==5.2.0
speedtest-cli==2.0.2
